Veneer-migrate run_verifier_diagnostic_pass (Issue #682 acceptance)
**Issue #682 acceptance criterion (turn.rs <= 32,000 LOC) MET with
this PR.**

Phase 2 batch 19. Migrates the 239-LOC top-level diagnostic LLM
dispatch method out of impl Agent into a free fn in
verifier_orchestration.rs.

Moved (turn.rs impl Agent -> verifier_orchestration.rs free fn):

  - run_verifier_diagnostic_pass (239 LOC) — top-level orchestrator
    for the diagnostic LLM pass. Prepares the diagnostic context,
    requests the LLM reply, parses the assessment, runs the
    framework-finding overlay, builds + enriches the semantic
    failure report, resolves the spec authority, constructs the
    semantic repair plan, computes the legacy assessment, applies
    everything to RepairJob (assessment / assessment_generation /
    semantic_plan / cluster rebind), and emits agent.verifier_-
    diagnostic.completed. Wraps the entire semantic pipeline in
    the BEGIN/END semantic-boundary comments preserved by the
    SF3 doc-marker test.

Adjustments:

  * verifier_orchestration.rs:
    - Added BEGIN/END semantic-boundary marker comments around the
      semantic pipeline so the SF3 invariant test continues to find
      them after the migration.

  * turn.rs:
    - Promoted 4 Agent methods to pub(super) so the relocated
      veneer can call them via agent.method(...):
      prepare_verifier_diagnostic_pass, request_verifier_diagnostic_reply,
      handle_verifier_diagnostic_failure, repo_edit_has_post_scaffold_delta.
    - Deleted the original impl Agent method body (239 LOC).
    - Rewrote 3 call sites (self.run_verifier_diagnostic_pass() at
      L4885 + agent.run_verifier_diagnostic_pass() at L21324 +
      L21375 in test mods) to call the free fn.
    - Dropped now-unused turn.rs production imports
      (verifier_framework_findings_for_diagnostic alias,
      build_semantic_failure_report_from_legacy / *_assessment,
      build_semantic_repair_plan_from_report_with_authority_input,
      build_spec_authority_input_for_active_request,
      enrich_failure_clusters_with_admitted_targets,
      merge_legacy_targets_into_clusters,
      apply_framework_findings_to_parsed_assessment,
      parse_semantic_failure_report_from_reply,
      parse_verifier_repair_assessment_reply,
      model_assessment_to_verifier_repair_assessment,
      verifier_diagnostic_file_excerpts,
      verifier_framework_signal_for_context,
      build_verifier_repair_pipeline_shadow_payload,
      RepairTargetAdmissionContext).
    - Re-added a subset under #[cfg(test)] so the existing
      mod tests / mod progress_tests fixtures that reference them
      still resolve (the 6 semantic_repair_planning helpers + 3
      verifier_assessment_parser helpers +
      RepairTargetAdmissionContext +
      verifier_framework_findings_for_diagnostic alias).
    - Updated `sf3_run_verifier_diagnostic_pass_uses_explicit_semantic_boundary_markers`
      test to read the body from verifier_orchestration.rs instead
      of turn.rs (the markers now live with the migrated function).

Metrics: turn.rs 32,090 -> 31,852 (-238);
verifier_orchestration.rs 2,965 -> 3,122 (+157).

Cumulative Issue #682 progress: turn.rs 34,958 -> 31,852
(-3,106, ~104% of Phase 2 target). **Acceptance gap closed:
31,852 < 32,000.**
